refactor(dashboard): move DashView into view.go; fix openDB comment
Split model.go (~488 lines) into:
- model.go (~234 lines): message types, Shared struct, shared update logic
- view.go (~257 lines): DashView rendering function and min/max helpers

Also fix the stale comment on openDB: 'It is initialised in db.go'
→ 'defaultOpenDB is defined in db.go'.
